[Fix] MigrationPlan 구현 테스트 및 파일 정리
