--- MC2-OneShot/Data/SwiftData/SwiftData.swift
@@ -1,13 +1,0 @@
-//
-//  SwiftData.swift
-//  MC2-OneShot
-//
-//  Created by 김민준 on 6/16/24.
-//
-
-import Foundation
-
-typealias Party = OneShotSchemaV2.Party
-typealias Step = OneShotSchemaV2.Step
-typealias Media = OneShotSchemaV2.Media
-typealias Member = OneShotSchemaV2.Member
