+++ MC2-OneShot/Domain/Entity/Media.swift
@@ -2,21 +2,9 @@
 //  Media.swift
 //  MC2-OneShot
 //
-//  Created by 김민준 on 6/1/24.
+//  Created by 김민준 on 6/16/24.
 //
 
 import Foundation
-import SwiftData
 
-// MARK: - Media
-
-//@Model
-//class Media: Identifiable {
-//    var fileData: Data
-//    var captureDate: Date
-//    
-//    init(fileData: Data, captureDate: Date) {
-//        self.fileData = fileData
-//        self.captureDate = captureDate
-//    }
-//}
+typealias Media = OneShotSchemaV2.Media
