+++ MC2-OneShot/Domain/Entity/Member.swift
@@ -2,21 +2,9 @@
 //  Member.swift
 //  MC2-OneShot
 //
-//  Created by 김민준 on 6/1/24.
+//  Created by 김민준 on 6/16/24.
 //
 
 import Foundation
-import SwiftData
 
-// MARK: - Member
-
-//@Model
-//class Member: Identifiable {
-//    let profileImageData: Data
-//    let captureDate: Date
-//    
-//    init(profileImageData: Data, captureDate: Date = .now) {
-//        self.profileImageData = profileImageData
-//        self.captureDate = captureDate
-//    }
-//}
+typealias Member = OneShotSchemaV2.Member
